登录、申请单、新添字段

+++ src/main/java/com/cmb/bankcheck/service/impl/CustomerServiceImpl.java
@@ -46,7 +46,7 @@ public class CustomerServiceImpl implements CustomerService {
         String key = apply.getProcessKey();
         String userId = apply.getUserId();
         String Xmtype = apply.getXmtype();
-        float amt = apply.getAmt();
+        double amt = apply.getAmt();
         System.out.println(key + ""+ Xmtype);
         if (key == null || userId == null || Xmtype == null || amt == 0){
             Message msg = new Message();
